+++ packages/integration-tests/tests/admin-user.test.ts
@@ -1,25 +1,25 @@
 import { HTTPError } from 'got';
 
-import { createUser } from '@/actions';
 import { getUser, getUsers, updateUser, deleteUser, updateUserPassword } from '@/api';
+import { createUserByAdmin } from '@/helpers';
 
 describe('admin console user management', () => {
   it('should create user successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
 
     const userDetails = await getUser(user.id);
     expect(userDetails.id).toBe(user.id);
   });
 
   it('should get user list successfully', async () => {
-    await createUser();
+    await createUserByAdmin();
     const users = await getUsers();
 
     expect(users.length).not.toBeLessThan(1);
   });
 
   it('should update userinfo successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
 
     const newUserData = {
       name: 'new name',
@@ -36,7 +36,7 @@ describe('admin console user management', () => {
   });
 
   it('should delete user successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
 
     const userEntity = await getUser(user.id);
     expect(userEntity).toMatchObject(user);
@@ -48,7 +48,7 @@ describe('admin console user management', () => {
   });
 
   it('should update user password successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
     const userEntity = await updateUserPassword(user.id, 'new_password');
     expect(userEntity).toMatchObject(user);
   });
