+++ packages/integration-tests/tests/me.test.ts
@@ -1,12 +1,12 @@
 import { ArbitraryObject, UserInfo, userInfoSelectFields } from '@logto/schemas';
 
-import { createUser } from '@/actions';
 import { api } from '@/api';
+import { createUserByAdmin } from '@/helpers';
 import { generatePassword } from '@/utils';
 
 describe('api `/me`', () => {
   it('should get user info successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
     const userInfo = await api
       .get(`me`, { headers: { 'development-user-id': user.id } })
       .json<UserInfo>();
@@ -19,7 +19,7 @@ describe('api `/me`', () => {
   });
 
   it('should get user custom data successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
     const customData = await api
       .get('me/custom-data', {
         headers: {
@@ -32,7 +32,7 @@ describe('api `/me`', () => {
   });
 
   it('should update user custom data successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
 
     const headers = Object.freeze({
       'development-user-id': user.id,
@@ -59,7 +59,7 @@ describe('api `/me`', () => {
   });
 
   it('should change user password successfully', async () => {
-    const user = await createUser();
+    const user = await createUserByAdmin();
     const password = generatePassword();
     const changePasswordResponse = await api.patch('me/password', {
       headers: {
